Use cylinders instead of STLs for wheel collision geometries
When using the STL meshes, there were lots of inconsistent contacts
between the wheels and the ground. This resulted in lots of small
"vibrations" of the robot due to unstable repulsive forces between
ground and wheels, which were measured by the simulated IMU. Using
cylinders as collision geometries fixes that.

Fixes #99.